requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.45.0
django-bulma==0.8.2
django-crispy-forms==1.12.0
django-crontab==0.7.1
django-debug-toolbar==3.2.2
django-environ==0.4.5
django-extensions==3.1.3
django-multiselectfield==0.1.12
django-paypal==1.1.1
django-redis-cache==2.1.0
django-rq==2.1.0
django-social-share==2.2.1
django-translation-manager==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
idna==3.2
image==1.5.33
img2pdf==0.4.1
joblib==1.0.1
lxml==4.6.3
nltk==3.6.3
oauthlib==3.1.1
paypal==1.2.5
pikepdf==3.1.0
Pillow==8.3.1
platformdirs==2.2.0
polib==1.1.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2021.1
PyVimeo==1.1.0
redis==3.5.3
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rq==1.9.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
textblob==0.15.3
tinydb==4.5.2
tqdm==4.62.3
tuspy==0.2.5
urllib3==1.26.6
whitenoise==5.3.0
virtualenv==20.7.1
django-heroku